Semi-correct NetDevice design. Worked on InMemDevice, now expanding for general ANetDevice solution.
The issue now dealing with is: when message is to be received, it checks if a connection from given sender exists. If not, message is ignored. But for messages informing about or confirming new connections, this connection doesn't exist yet, so the message is ignored. Fix for this will need some larger restructuring of existing code, hence this commit ;)