Added support for conditional SCL processing.
Assigment and command call lines can be started with '?X' where X is either offset of the flag bit to test on or a ASCII alt name (like '!' as If flag, '~' as else flag, and so on).
The support is limited to only this most basic scenario. Test for binary complement is not supported. Flags need to be set manually inside of called command. This change also limits number of potential arguments per-command and since extra args isn't currently supported, this can be limiting change.